ml-models/requirements.txt
joblib
scikit-learn
Pillow
numpy
opencv-python
tensorflow
googlesearch-python
beautifulsoup4
h5py
